requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
openai>=1.30.0
google-genai>=0.3.0
slowapi>=0.1.9
pytest>=8.2.0
hypothesis>=6.100.0
httpx>=0.27.0
